src/main/webapp/app/entities/estimate-basis/detail/estimate-basis-detail.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { ActivatedRoute } from '@angular/router';
import { of } from 'rxjs';

import { DataUtils } from 'app/core/util/data-util.service';

import { EstimateBasisDetailComponent } from './estimate-basis-detail.component';

describe('EstimateBasis Management Detail Component', () => {
  let comp: EstimateBasisDetailComponent;
  let fixture: ComponentFixture<EstimateBasisDetailComponent>;
  let dataUtils: DataUtils;

  beforeEach(() => {
    TestBed.configureTestingModule({
      declarations: [EstimateBasisDetailComponent],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: { data: of({ estimateBasis: { id: 123 } }) },
        },
      ],
    })
      .overrideTemplate(EstimateBasisDetailComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(EstimateBasisDetailComponent);
    comp = fixture.componentInstance;
    dataUtils = TestBed.inject(DataUtils);
    jest.spyOn(window, 'open').mockImplementation(() => null);
  });

  describe('OnInit', () => {
    it('Should load estimateBasis on init', () => {
      // WHEN
      comp.ngOnInit();

      // THEN
      expect(comp.estimateBasis).toEqual(expect.objectContaining({ id: 123 }));
    });
  });

  describe('byteSize', () => {
    it('Should call byteSize from DataUtils', () => {
      // GIVEN
      jest.spyOn(dataUtils, 'byteSize');
      const fakeBase64 = 'fake base64';

      // WHEN
      comp.byteSize(fakeBase64);

      // THEN
      expect(dataUtils.byteSize).toBeCalledWith(fakeBase64);
    });
  });

  describe('openFile', () => {
    it('Should call openFile from DataUtils', () => {
      const newWindow = { ...window };
      newWindow.document.write = jest.fn();
      window.open = jest.fn(() => newWindow);
      window.onload = jest.fn(() => newWindow);
      window.URL.createObjectURL = jest.fn();
      // GIVEN
      jest.spyOn(dataUtils, 'openFile');
      const fakeContentType = 'fake content type';
      const fakeBase64 = 'fake base64';

      // WHEN
      comp.openFile(fakeBase64, fakeContentType);

      // THEN
      expect(dataUtils.openFile).toBeCalledWith(fakeBase64, fakeContentType);
    });
  });
});
